Debug for SYMGS+SPMV fusion
Previous implementation of SYMGS+SPMV was using x_old to calculate the spmv (L * old_x + (D+U) * new_x), It was totally wrong due to the expression is equal to the value of source (i.e., b). The mathematical proof is that the backward induction is exact (U+D)x’’+Lx’=b. 

Now we ensure all x are new, and we calculate Ax by (L+D+U)x = b-Lx_pre+Lx to reuse the data calculated in forward sweep.